Increase organization max seat size from 30k to 2b (#1274)
* Increase organization max seat size from 30k to 2b

* PR review. Do not modify unless state matches expected

+++ src/Core/Models/Business/OrganizationLicense.cs
@@ -99,7 +99,7 @@ namespace Bit.Core.Models.Business
         public bool Enabled { get; set; }
         public string Plan { get; set; }
         public PlanType PlanType { get; set; }
-        public short? Seats { get; set; }
+        public int? Seats { get; set; }
         public short? MaxCollections { get; set; }
         public bool UsePolicies { get; set; }
         public bool UseSso { get; set; }
